+++ livestock-tracker.abstractions/Services/Feed/IFeedingTransactionCrudService.cs
@@ -6,7 +6,7 @@ namespace LivestockTracker.Abstractions.Services.Feed
     /// Provides create, read and update and deletion services for feeding transactions.
     /// Also provides paged record retrieval services.
     /// </summary>
-    public interface IFeedingTransactionCrudService : ICrudAsyncService<IFeedingTransaction, int>, IPagedFetchAsyncService<IFeedingTransaction>
+    public interface IFeedingTransactionCrudService : ICrudAsyncService<IFeedingTransaction, long>, IPagedFetchAsyncService<IFeedingTransaction>
     {
     }
 }
