+++ livestock-tracker.abstractions/Services/Medical/IMedicalTransactionCrudService.cs
@@ -5,7 +5,7 @@ namespace LivestockTracker.Abstractions.Services.Medical
     /// <summary>
     /// Provides create, read and update and deletion services for medical transactions.
     /// </summary>
-    public interface IMedicalTransactionCrudService : ICrudAsyncService<IMedicalTransaction, int>
+    public interface IMedicalTransactionCrudService : ICrudAsyncService<IMedicalTransaction, long>
     {
     }
 }
