+++ livestock-tracker.base/EntityNotFoundException.cs
@@ -4,7 +4,7 @@ namespace LivestockTracker
 {
   public class EntityNotFoundException<T> : Exception
   {
-    public EntityNotFoundException(int id) : base($"{typeof(T).Name} with id {id} not found.")
+    public EntityNotFoundException(long id) : base($"{typeof(T).Name} with id {id} not found.")
     {
     }
   }
